# Tune parser for query with lots of nested fields