--- a/lib/graphql/language/lexer.rl
+++ b/lib/graphql/language/lexer.rl
@@ -67,14 +67,14 @@
   main := |*
     INT           => { emit(:INT, ts, te, meta) };
     FLOAT         => { emit(:FLOAT, ts, te, meta) };
-    ON            => { emit(:ON, ts, te, meta) };
-    FRAGMENT      => { emit(:FRAGMENT, ts, te, meta) };
-    TRUE          => { emit(:TRUE, ts, te, meta) };
-    FALSE         => { emit(:FALSE, ts, te, meta) };
-    NULL          => { emit(:NULL, ts, te, meta) };
-    QUERY         => { emit(:QUERY, ts, te, meta) };
-    MUTATION      => { emit(:MUTATION, ts, te, meta) };
-    SUBSCRIPTION  => { emit(:SUBSCRIPTION, ts, te, meta) };
+    ON            => { emit(:ON, ts, te, meta, "on") };
+    FRAGMENT      => { emit(:FRAGMENT, ts, te, meta, "fragment") };
+    TRUE          => { emit(:TRUE, ts, te, meta, "true") };
+    FALSE         => { emit(:FALSE, ts, te, meta, "false") };
+    NULL          => { emit(:NULL, ts, te, meta, "null") };
+    QUERY         => { emit(:QUERY, ts, te, meta, "query") };
+    MUTATION      => { emit(:MUTATION, ts, te, meta, "mutation") };
+    SUBSCRIPTION  => { emit(:SUBSCRIPTION, ts, te, meta, "subscription") };
     SCHEMA        => { emit(:SCHEMA, ts, te, meta) };
     SCALAR        => { emit(:SCALAR, ts, te, meta) };
     TYPE          => { emit(:TYPE, ts, te, meta) };
@@ -85,22 +85,22 @@
     ENUM          => { emit(:ENUM, ts, te, meta) };
     INPUT         => { emit(:INPUT, ts, te, meta) };
     DIRECTIVE     => { emit(:DIRECTIVE, ts, te, meta) };
-    RCURLY        => { emit(:RCURLY, ts, te, meta) };
-    LCURLY        => { emit(:LCURLY, ts, te, meta) };
-    RPAREN        => { emit(:RPAREN, ts, te, meta) };
-    LPAREN        => { emit(:LPAREN, ts, te, meta) };
-    RBRACKET      => { emit(:RBRACKET, ts, te, meta) };
-    LBRACKET      => { emit(:LBRACKET, ts, te, meta) };
-    COLON         => { emit(:COLON, ts, te, meta) };
+    RCURLY        => { emit(:RCURLY, ts, te, meta, "}") };
+    LCURLY        => { emit(:LCURLY, ts, te, meta, "{") };
+    RPAREN        => { emit(:RPAREN, ts, te, meta, ")") };
+    LPAREN        => { emit(:LPAREN, ts, te, meta, "(")};
+    RBRACKET      => { emit(:RBRACKET, ts, te, meta, "]") };
+    LBRACKET      => { emit(:LBRACKET, ts, te, meta, "[") };
+    COLON         => { emit(:COLON, ts, te, meta, ":") };
     QUOTED_STRING => { emit_string(ts, te, meta, block: false) };
     BLOCK_STRING  => { emit_string(ts, te, meta, block: true) };
-    VAR_SIGN      => { emit(:VAR_SIGN, ts, te, meta) };
-    DIR_SIGN      => { emit(:DIR_SIGN, ts, te, meta) };
-    ELLIPSIS      => { emit(:ELLIPSIS, ts, te, meta) };
-    EQUALS        => { emit(:EQUALS, ts, te, meta) };
-    BANG          => { emit(:BANG, ts, te, meta) };
-    PIPE          => { emit(:PIPE, ts, te, meta) };
-    AMP           => { emit(:AMP, ts, te, meta) };
+    VAR_SIGN      => { emit(:VAR_SIGN, ts, te, meta, "$") };
+    DIR_SIGN      => { emit(:DIR_SIGN, ts, te, meta, "@") };
+    ELLIPSIS      => { emit(:ELLIPSIS, ts, te, meta, "...") };
+    EQUALS        => { emit(:EQUALS, ts, te, meta, "=") };
+    BANG          => { emit(:BANG, ts, te, meta, "!") };
+    PIPE          => { emit(:PIPE, ts, te, meta, "|") };
+    AMP           => { emit(:AMP, ts, te, meta, "&") };
     IDENTIFIER    => { emit(:IDENTIFIER, ts, te, meta) };
     COMMENT       => { record_comment(ts, te, meta) };
 
@@ -163,11 +163,11 @@ module GraphQL
 
       def self.record_comment(ts, te, meta)
         token = GraphQL::Language::Token.new(
-          name: :COMMENT,
-          value: meta[:data][ts, te - ts].pack(PACK_DIRECTIVE).force_encoding(UTF_8_ENCODING),
-          line: meta[:line],
-          col: meta[:col],
-          prev_token: meta[:previous_token],
+          :COMMENT,
+          meta[:data][ts, te - ts].pack(PACK_DIRECTIVE).force_encoding(UTF_8_ENCODING),
+          meta[:line],
+          meta[:col],
+          meta[:previous_token],
         )
 
         meta[:previous_token] = token
@@ -175,13 +175,14 @@ module GraphQL
         meta[:col] += te - ts
       end
 
-      def self.emit(token_name, ts, te, meta)
+      def self.emit(token_name, ts, te, meta, token_value = nil)
+        token_value ||= meta[:data][ts, te - ts].pack(PACK_DIRECTIVE).force_encoding(UTF_8_ENCODING)
         meta[:tokens] << token = GraphQL::Language::Token.new(
-          name: token_name,
-          value: meta[:data][ts, te - ts].pack(PACK_DIRECTIVE).force_encoding(UTF_8_ENCODING),
-          line: meta[:line],
-          col: meta[:col],
-          prev_token: meta[:previous_token],
+          token_name,
+          token_value,
+          meta[:line],
+          meta[:col],
+          meta[:previous_token],
         )
         meta[:previous_token] = token
         # Bump the column counter for the next token
@@ -220,30 +221,30 @@ module GraphQL
         # (It's faster: https://bugs.ruby-lang.org/issues/8110)
         if !value.valid_encoding? || value !~ VALID_STRING
           meta[:tokens] << token = GraphQL::Language::Token.new(
-            name: :BAD_UNICODE_ESCAPE,
-            value: value,
-            line: meta[:line],
-            col: meta[:col],
-            prev_token: meta[:previous_token],
+            :BAD_UNICODE_ESCAPE,
+            value,
+            meta[:line],
+            meta[:col],
+            meta[:previous_token],
           )
         else
           replace_escaped_characters_in_place(value)
 
           if !value.valid_encoding?
             meta[:tokens] << token = GraphQL::Language::Token.new(
-              name: :BAD_UNICODE_ESCAPE,
-              value: value,
-              line: meta[:line],
-              col: meta[:col],
-              prev_token: meta[:previous_token],
+              :BAD_UNICODE_ESCAPE,
+              value,
+              meta[:line],
+              meta[:col],
+              meta[:previous_token],
             )
           else
             meta[:tokens] << token = GraphQL::Language::Token.new(
-              name: :STRING,
-              value: value,
-              line: meta[:line],
-              col: meta[:col],
-              prev_token: meta[:previous_token],
+              :STRING,
+              value,
+              meta[:line],
+              meta[:col],
+              meta[:previous_token],
             )
           end
         end