@DmitryTsepelev 2019-10-30
Wow, 2x memory reduce is a huge win!