@rmosolgo 2019-10-30
I don't know how much it really matters in a production setting, but I enjoyed the hunt :P Thanks again for sharing the benchmarks. It looks like it was a great talk too, even though all I could do was read the slides!

I also have a suggestion for improving the runtime benchmark: 

```diff 
  class GraphqlSchema < GraphQL::Schema
    query QueryType
+   use GraphQL::Execution::Interpreter
+   use GraphQL::Analysis::AST
  end
# And change the warmup: 

-   GraphqlSchema.graphql_definition 
+   GraphqlSchema.execute(query, variables: { id: "1" })
```

```
Total allocated: 69.65 kB (704 objects)
Total retained:  0 B (0 objects)

allocated memory by gem
-----------------------------------
  53.78 kB  graphql-ruby/lib
   7.19 kB  other
   5.81 kB  ostruct
   2.33 kB  racc
  544.00 B  forwardable
```

About a 40% improvement! I'm hoping to make `Interpreter` the default in 1.12.