@DmitryTsepelev 2019-10-30
> It looks like it was a great talk too, even though all I could do was read the slides!

Thank you soo much! I plan to give it at least once in English somewhere 🙂I've updated the benchmark and got almost the same performance boost, will re-run it again as soon as this PR is merged and released